requirements.txt
langchain
langchain-ibm
langchain-community 
ibm-watsonx-ai 
unstructured
chromadb 
python-dotenv=
beautifulsoup4